Add proper parsing to DMs
